Redirects to a non-existent route
I was having some issues with the last part of Chapter 7, and this fix seemed to address them. I think the problem was an empty/incomplete form redirecting to a non-existent route. Routes of the form "/location/' + locationid + '/reviews/" belong to the api_server not the app_server. Please let me know what you think. Thank you,